Restructured namespaces
Removed deprecated comment
