+++ ObjectPrinting/Solved/Extensions/ObjectExtensions.cs
@@ -1,4 +1,4 @@
-namespace ObjectPrinting.Solved
+namespace ObjectPrinting.Solved.Extensions
 {
     public static class ObjectExtensions
     {
