+++ ObjectPrinting/Solved/Extensions/PropertyPrintingConfigExtensions.cs
@@ -1,6 +1,7 @@
 using System;
+using ObjectPrinting.Solved.PrintingConfiguration;
 
-namespace ObjectPrinting.Solved
+namespace ObjectPrinting.Solved.Extensions
 {
     public static class PropertyPrintingConfigExtensions
     {
@@ -9,10 +10,10 @@ namespace ObjectPrinting.Solved
             return config(ObjectPrinter.For<T>()).PrintToString(obj);
         }
 
-        public static PrintingConfig<TOwner> TrimmedToLength<TOwner>(this PropertyPrintingConfig<TOwner, string> propConfig, int maxLen)
+        public static PrintingConfig<TOwner> TrimmedToLength<TOwner>(
+            this PropertyPrintingConfig<TOwner, string> propConfig, int maxLen)
         {
             return ((IPropertyPrintingConfig<TOwner, string>)propConfig).ParentConfig;
         }
-
     }
 }
