+++ ObjectPrinting/Solved/Extensions/TypeExtensions.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
 
-namespace ObjectPrinting.Solved
+namespace ObjectPrinting.Solved.Extensions
 {
     public static class TypeExtensions
     {
